requirements.txt
fastapi==0.110.1
pillow==10.3.0
torch==2.0.0
torchvision==0.15.1
transformers==4.40.2
uvicorn==0.27.0
numpy<2
